fix: address canary replay review feedback
