+++ tests/test_agent/test_task_runtime.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from datetime import date, datetime
-from pathlib import Path
 from types import ModuleType, SimpleNamespace
 import sys
